+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2019-09-04
+### Added
+- Channel add_new_consumer method to add a new consumer with filters
+- Channel get_consumer_from_filters to get a list of consumers that match with filters
+
+### Changed
+- Channel new_consumer method can handle a consumer filters dict
+- Channel __add_new_consumer_and_run to use consumer filters and not consumer name
+
 ## [1.1.14] - 2019-08-29
 ### Added
 - Tests